requirements.txt
Django>=5.0,<5.3
pandas>=2.0
openpyxl>=3.1
xhtml2pdf>=0.2.15

django-simple-history
python-dotenv
